refactor: move list to houseQueryService
